+++ builder_pkgs/mockito/lib/mirrors.dart
@@ -1,0 +1,2 @@
+export 'mockito.dart';
+export 'src/spy.dart' show spy;