+++ builder_pkgs/mockito/lib/mockito_no_mirrors.dart
@@ -1,4 +1,4 @@
-@Deprecated('Import "mockito.dart" instead.')
+@Deprecated('Import "package:mockito/mockito.dart" instead.')
 library mockito.mockito_no_mirrors;
 
 export 'mockito.dart';